+++ src/components/ui/CommandPalette.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import { Dialog, DialogContent } from "@/ui/dialog";
-import { dashboardRoutes } from "@/routes";
+import { dashboardRoutes } from "@/routes/index";
 
 interface CommandPaletteProps {
   open?: boolean;
